Function ic_crypto_internal_bls12381_serde_miracl::miracl_g1_from_bytes

pub fn miracl_g1_from_bytes(bytes: &[u8; 48]) -> Result<ECP, ()>

Parses a G1 in a standard, library-independent form to a MIRACL ECP.

Also verifies that the point is in the correct prime order subgroup.

Errors

Err(()) if

The point is not in the correct prime order subgroup.

The point is encoded in UNCOMPRESSED form

The point's x-coordinate is non-canonical (i.e. greater than the field modulus)

The point's x-coordinate is infinity, but the INFINITY flag is not set

The INFINITY flag is set, but the encoded x-coordinate is not all-zeroes